Еще правки по комментариям.
