+++ MT.Web/MT.Utility/MT.Utility.csproj
@@ -111,6 +111,7 @@
     <Compile Include="OtherTools\CustomUtilityExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringExtensions.cs" />
+    <Compile Include="Validation\ServerValidationService.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
