+++ MT.Web/MT.Web/Controllers/PersonalCabinet/PersonalCabinetController.cs
@@ -6,6 +6,7 @@ using MT.DomainLogic.PersonalCabinet;
 using MT.ModelEntities.Entities;
 using MT.Utility.Json;
 using MT.Utility.OtherTools;
+using MT.Utility.Validation;
 using MT.Web.ViewModels;
 
 namespace MT.Web.Controllers.PersonalCabinet
@@ -57,7 +58,7 @@ namespace MT.Web.Controllers.PersonalCabinet
 
             if (!ModelState.IsValid)
             {
-                return ErrorModel.CreateErrorModelViaModelState(ModelState);
+                return ServerValidationService.CreateErrorModelViaModelState(ModelState);
             }
 
             var userMail = Request.Cookies.GetUserMailFromFormsAuthentication();
